sales-sync-backend/requirements.txt
PyMySQL
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
marshmallow==3.19.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.2
mysqlclient==2.2.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
geopy==2.3.0
Pillow==10.0.0
flask-cors==4.0.0
gunicorn==20.1.0